Use more responsive <C-V> mappings
The ones from paste.vim work, but because they depend on typing a couple
of characters to position the cursor, they inevitably cause the screen
to flash something for a moment. The new mappings make use of a function
call the changes the "+ register to be characterwise, so that more
standard pasting commands can be used ("+gP in normal mode, <C-R><C-O>+
in insert mode). The "-c approach for visual mode makes the screen
flicker, so I've rolled out a completely different approach that I'm
fairly sure works.