Continue even after failing
ln fails for not very useful reasons like the file already existing, in
which case it's better to just fix those single problems rather than
making the whole script fail. i.e. the commands are mostly independent
and can be run in parallel so it doesn't really make sense to abort
after a single fail.