Remove oldfugitive
I don't use this anymore and was just leaving it in out of laziness and
optionality. But then I realized that I hadn't properly removed the
:Gblame command when I created oldfugitive, so that when I thought I was
using the latest :Gblame (which doesn't exist anymore in the latest
fugitive; it's called :G blame now), in fact I was using the years-old
blame that doesn't properly work with :GBrowse. It's not worth the
trouble to fix oldfugitive when I don't even use it, so out it goes.