Set viminfo
For some reason, on Fedora, /etc/vimrc sets viminfo to have '20, so that
only the last 20 files have their location remembered. I've frequently
run into situations where I know I opened a file recently, but when I
edit it I am left at the start of the file rather than wherever I was
editing last, and it is kind of annoying to have to find where I was.
But if I am changing the value of one of the viminfo parameters anyway,
I might as well just load up the Neovim default.