kitty: don't remember window size
It turns out that i tend to do different things in the morning vs night,
so the size i like to have at night (larger window for programming)
tends to be a poor fit for morning (smaller window for random tasks). So
resetting the window size to the "morning size" is better than
remembering what size i was using the previous night. (I shut down my
laptop every night.)