Tweak theme
I got annoyed that opening certain files in neovim was causing the
screen to flash black when using a light terminal, and didn't quite like
the default neovim theme because all the colors blend together.

Thinking more about it, the only reason I preferred the default neovim
theme compared to the default vim theme is that the spelling highlight
on the default vim dark theme is terrible (it makes the text
unreadable). But on kitty it is possible to get the undercurls like on
the default neovim theme. So I'm switching back to the default default
vim dark theme + undercurls for the spelling.

Also, I've commented out the let &filetype = &filetype because that
still seems to cause a weird flash (but not the full screen black from
before, instead only the text color seems to change).