Use `silent! set` more
I realized that for simple options like tags or belloff or whatever,
where I'm not trying to set a different value between +eval and -eval
versions of Vim, it doesn't really make sense to do the logic for both
versions. Doing just the `silent!` version makes it work on both kinds
of Vim.

For things like mappings, where I do a crappy fallback version for
-eval, and a nicer version for +eval, it does still make sense to do the
logic on both.